signing 수정 / 불필요 주석 삭제
